add base controller for DB storage
