+++ core/components/minishop2/handlers/storage/db/orderdbhandler.class.php
@@ -1,28 +1,12 @@
 <?php
 
-class OrderDBHandler
-{
-    protected $modx;
-    protected $ctx = 'web';
-    protected $ms2;
-    /**
-     * @var msOrder $msOrder
-     */
-    protected $msOrder;
-    /**
-     * @var msOrderAddress $address
-     */
-    protected $address;
+require_once 'baseDBController.php';
 
+class OrderDBHandler extends BaseDBController
+{
     public function __construct(modX $modx, miniShop2 $ms2)
     {
-        $this->modx = $modx;
-        $this->ms2 = $ms2;
-    }
-
-    public function setContext($ctx)
-    {
-        $this->ctx = $ctx;
+        parent::__construct($modx, $ms2);
     }
 
     public function get()
@@ -113,7 +97,6 @@ class OrderDBHandler
             'user_id' => $data['user_id'],
             'updatedon' => time(),
         ]);
-        $msOrder->addOne($address);
         $msOrder->save();
         return $msOrder;
     }
@@ -126,22 +109,5 @@ class OrderDBHandler
         $this->msOrder->set('delivery_cost', $delivery_cost);
         $this->msOrder->set('cost', $cost);
         $this->msOrder->save();
-    }
-
-    private function getStorageOrder()
-    {
-        $where = ['status' => 999];
-        $user_id = $this->modx->getLoginUserID($this->ctx);
-        if ($user_id > 0) {
-            //TODO реализовать вопрос склеивания корзин анонима и залогиненного юзера
-            $where['user_id'] = $user_id;
-        } else {
-            $where['session_id'] = session_id();
-        }
-        $q = $this->modx->newQuery('msOrder');
-        $q->sortby('updatedon', 'DESC');
-        $q->where($where);
-        /** @var msOrder $msOrder */
-        return $this->modx->getObject('msOrder', $q);
     }
 }
